requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-JWT-Extended==4.6.0
